backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-multipart
python-dotenv
ffmpeg-python
openai

PyYAML
pydantic-settings
